Deduplicate zone parsing code
